# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dayfi API
  version: 1.0.0
  description: |
    Dayfi backend API — unified **USD** ledger with optional **NGN** local spend.

    - Human-readable reference: [API.md](./API.md)
    - Architecture: [PAYMENTS_ARCHITECTURE.md](./PAYMENTS_ARCHITECTURE.md)

    **Base path:** `/api/v1`
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: Local development
  - url: https://api.dayfi.co/api/v1
    description: Production (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
        code:
          type: integer
        data:
          nullable: true
    ApiError:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
        message:
          type: string
...
    PayoutQuote:
      type: object
      properties:
        sendAmountUsd:
          type: number
        targetCurrency:
          type: string
        exchangeRate:
          type: number
        beneficiaryReceives:
          type: number
        feeUsd:
          type: number
        estimatedMinutes:
          type: integer
          example: 30
...
    SwapRequest:
      type: object
      required: [fromCurrency, toCurrency, amount]
      properties:
        fromCurrency:
          type: string
          enum: [USD, NGN, EUR, GBP, CAD]
        toCurrency:
          type: string
          enum: [USD, NGN, EUR, GBP, CAD]
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
      example:
        fromCurrency: USD
        toCurrency: NGN
        amount: 100
    ExchangeRateUpsert:
      type: object
      required: [baseCurrency, targetCurrency, rate]
      properties:
        baseCurrency:
          type: string
          enum: [USD, NGN, EUR, GBP, CAD]
        targetCurrency:
          type: string
          enum: [USD, NGN, EUR, GBP, CAD]
        rate:
          type: number
          exclusiveMinimum: true
        source:
          type: string
          default: manual
    BankTransferRequest:
      type: object
      required:
        - amount
        - accountNumber
        - bankCode
        - bankName
        - accountName
        - fee
        - pin
      properties:
        amount:
          type: number
        accountNumber:
          type: string
        bankCode:
          type: string
        bankName:
          type: string
        accountName:
          type: string
        fee:
          type: number
        pin:
          type: string
        spendCurrency:
          type: string
          enum: [USD, NGN]
          default: NGN
          description: |
            NGN — legacy Flutterwave payout from NGN wallet.
            USD — requires Grey (400 if not configured).
...
    YellowCardCollectionRequest:
      type: object
      required:
        - amount
        - currency
        - channelId
        - channelName
        - country
        - recipient
        - source
      properties:
        amount:
          type: number
        currency:
          type: string
        channelId:
          type: string
        channelName:
          type: string
        country:
          type: string
        reason:
          type: string
        recipient:
          type: object
          required: [name, country, phone, address, dob, email, idNumber, idType]
        source:
          type: object
          required: [accountNumber, accountType, networkId]
...
security:
  - bearerAuth: []
paths:
...
  /payments/wallets/swap:
    post:
      tags: [FX]
      summary: Swap between ledger wallets (e.g. USD → NGN)
      operationId: swapCurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapRequest'
      responses:
        '200':
          description: Swap completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400':
          description: Insufficient balance or missing rate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /payments/exchange-rate:
    get:
      tags: [FX]
      summary: Get platform FX rate
      operationId: getExchangeRate
      security:
        - bearerAuth: []
      parameters:
        - name: baseCurrency
          in: query
          required: true
          schema:
            type: string
            enum: [USD, NGN, EUR, GBP, CAD]
        - name: targetCurrency
          in: query
          required: true
          schema:
            type: string
            enum: [USD, NGN, EUR, GBP, CAD]
      responses:
        '200':
          description: Rate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
    post:
      tags: [FX]
      summary: Upsert platform FX rate (ops)
      operationId: createExchangeRate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRateUpsert'
      responses:
        '200':
          description: Rate saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
...
  /payments/create-collections:
    post:
      tags: [Inflows, Yellow Card]
      summary: Create Yellow Card collection
      description: |
        On `COLLECTION.COMPLETE` webhook, funds credit **USD** after FX.
      operationId: createCollectionRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YellowCardCollectionRequest'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
...
  /payments/send/quote:
    get:
      tags: [Outflows]
      summary: Payout quote (rate, fees, beneficiary amount)
      description: |
        Journey 3 — Use before send confirmation. Configure USD→target via exchange-rate.
      operationId: getPayoutQuote
      parameters:
        - name: amountUsd
          in: query
          required: true
          schema:
            type: number
        - name: targetCurrency
          in: query
          required: true
          schema:
            type: string
            example: GHS
        - name: feeUsd
          in: query
          schema:
            type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PayoutQuote'
...
  /payments/bank-transfer:
    post:
      tags: [Outflows, Legacy]
      summary: Nigeria bank transfer (Flutterwave)
      description: |
        Default `spendCurrency` is NGN. USD requires Grey configuration.
      operationId: bankTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankTransferRequest'
      responses:
        '200':
          description: Transfer initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400':
          description: Grey required for USD spend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...